backend/requirements-minimal.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.0
email-validator==2.1.0
PyPDF2==3.0.1
langchain-core==0.2.15
langchain-google-genai==1.0.7
scikit-learn==1.3.0
numpy==1.24.3